Download the forms archive in addition to checking the version (#9)
* Download the forms archive in addition to checking the version

We have a few reports from the mailing list (e.g. [this one](https://groups.google.com/g/pat-users/c/haiP_jGK9qg/) of errors when updating forms, sometimes in the form of timeouts. Short of network problems (user reports that's not the case) or MS OneDrive outage (possible but shouldn't happen often), there's another possibility in the design of the system.

Because this repo's GitHub action triggers scraping once an hour, there's a time period during which an update on the Winlink site could leave us with a stale download URL, which could cause 404s or timeouts. To give us more control, this PR will start downloading the forms archive and serving it through Github Pages.

This PR does not yet change the archive URL published in the `latest` document. Instead, it logs the URL and commits the forms archive so we can test and make sure the change works before making it live.

* Check the Zip archive for integrity