Don't output Generated timestamp
Maybe only my OCD, but it seems excessive to force an update (and commit) only to update the timestamp. We can use the Actions status to see the same information.

We might need to modify the Actions script not to fail when git commit/push commands see no changes though.

What do you think, @xylo04?